got the augmented reality (sceneform) app functionality to work
